@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "poetry"
-version = "1.1.13"
+version = "1.1.14"
 description = "Python dependency management and packaging made easy."
 authors = [
     "Sébastien Eustace <sebastien@eustace.io>"
@@ -22,7 +22,7 @@ classifiers = [
 
 # Requirements
 [tool.poetry.dependencies]
-python = "~2.7 || ^3.5"
+python = "^3.5"
 
 poetry-core = "~1.0.7"
 cleo = "^0.8.1"
@@ -74,11 +74,11 @@ pytest-mock = "^1.9"
 pre-commit = { version = "^2.6", python = "^3.6.1" }
 tox = "^3.0"
 pytest-sugar = "^0.9.2"
-httpretty = "^0.9.6"
+httpretty = "^1.0.3"
 # We need to restrict the version of urllib3 to avoid
 # httpretty breaking. This is fixed in httpretty >= 1.0.3
 # but it's not compatible with Python 2.7 and 3.5.
-urllib3 = "~1.25.10"
+urllib3 = "~1.26.9"
 
 [tool.poetry.scripts]
 poetry = "poetry.console:main"
