@@ -1,8 +1,11 @@
 #!/usr/bin/env nix-shell
 #!nix-shell -i python3 -p python39 python39.pkgs.packaging python39.pkgs.beautifulsoup4 python39.pkgs.requests
+# mirrored in ./default.nix
 from packaging import version
 from bs4 import BeautifulSoup
 import re, requests, json
+import os, sys
+from pathlib import Path
 
 URL = "https://downloads.asterisk.org/pub/telephony/asterisk"
 
@@ -23,8 +26,10 @@ for mv in major_versions.keys():
         "sha256": sha
     }
 
+versions_path = Path(sys.argv[0]).parent / "versions.json"
+
 try:
-    with open("versions.json", "r") as in_file:
+    with open(versions_path, "r") as in_file:
         in_data = json.loads(in_file.read())
         for v in in_data.keys():
             print(v + ":", in_data[v]["version"], "->", out[v]["version"])
@@ -32,5 +37,5 @@ except:
     # nice to have for the PR, not a requirement
     pass
 
-with open("versions.json", "w") as out_file:
+with open(versions_path, "w") as out_file:
     out_file.write(json.dumps(out, sort_keys=True, indent=2) + "\n")
